# swagger/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Adventure APIs",
    "description": "API for admin/managing and client/viewing",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/trip_diary/create": {
      "post": {
        "tags": [
          "trip_diary"
        ],
        "description": "add trip diary",
        "operationId": "create",
        "parameters": [
          {
            "name": "token",
            "in": "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_trip",
            "in": "formData",
            "description": "id_trip",
            "required": false,
            "type": "string"
          },
          {
            "name": "title",
            "in": "formData",
            "description": "title",
            "required": false,
            "type": "string"
          },
          {
            "name": "content",
            "in": "formData",
            "description": "content",
            "required": false,
            "type": "string"
          },
          {
            "name": "image_description",
            "in": "formData",
            "description": "image and description. Example : [{'url':'https...','description':'aaa',{'url':'https...','description':'bbb'}]",
            "required": false,
            "type": "string"
          },
          {
            "name": "detail_diary",
            "in": "formData",
            "description": "detail_diary. Example: [{\"date\":\"2016-10-05T04:27:58.144Z\",\"title\":\"This's title\",\"content\":\"This's content\"},{}]",
            "required": false,
            "type": "string"
          },
          {
            "name": "permission",
            "in": "formData",
            "description": "permission. Example: 1: Only me, 2: Member in trip, 3: Public",
            "required": true,
            "type": "integer"
          },
          {
            "name": "type",
            "in": "formData",
            "description": "type. Example: 1: Diary in trip, 2: Diary on timelines, 3: Diary in group",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Redis error; -2: DB error; -3: Token is not found; -4: Trip is not found",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "id_trip": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "owner": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "title": {
                      "type": "string",
                      "example": "test diary"
                    },
                    "content": {
                      "type": "string",
                      "example": "this's content"
                    },
                    "images": {
                      "type": "array",
                      "items": {
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "5769097529f87b5c271728e5"
                          },
                          "image": {
                            "type": "string",
                            "example": ""
                          },
                          "description": {
                            "type": "string",
                            "example": "test description"
                          }
                        }
                      }
                    },
                    "detail_diary": {
                      "type": "array",
                      "items": {
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "5769097529f87b5c271728e5"
                          },
                          "date": {
                            "type": "date-time",
                            "example": "2016-10-05T04:27:58.144Z"
                          },
                          "title": {
                            "type": "string",
                            "example": "This's title"
                          },
                          "content": {
                            "type": "string",
                            "example": "This's content"
                          }
                        }
                      }
                    },
                    "permission": {
                      "type": "number",
                      "example": 1
                    },
                    "type": {
                      "type": "number",
                      "example": 1
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
    "/api/trip_diary/browse": {
      "post": {
        "tags": [
          "trip_diary"
        ],
        "description": "browse trip diary",
        "operationId": "browse",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_trip",
            "in": "query",
            "description": "id_trip",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page to pagination",
            "required": false,
            "type": "number"
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "item per page for pagination",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Redis error; -2: DB error; -3: Token is not found; -4: Trip is not found; -5: Trip diary is not found",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "owner": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "5769097529f87b5c271728e5"
                        },
                        "first_name": {
                          "type": "string",
                          "example": "Brucelee Thanh"
                        },
                        "last_name": {
                          "type": "string",
                          "example": "Brucelee Thanh"
                        },
                        "avatar": {
                          "type": "string",
                          "example": ""
                        }
                      }
                    },
                    "title": {
                      "type": "string",
                      "example": "this's title"
                    },
                    "images": {
                      "type": "array",
                      "items": {
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "5769097529f87b5c271728e5"
                          },
                          "image": {
                            "type": "string",
                            "example": ""
                          },
                          "description": {
                            "type": "string",
                            "example": "test description"
                          }
                        }
                      }
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
    "/api/trip_diary/detail": {
      "post": {
        "tags": [
          "trip_diary"
        ],
        "description": "get detail of trip diary",
        "operationId": "detail",
        "parameters": [
          {
            "name": "token",
            "in": "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_trip_diary",
            "in": "formData",
            "description": "id_trip",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Redis error; -2: DB error; -3: Token is not found; -4: Trip is not found",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "id_trip": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "owner": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "title": {
                      "type": "string",
                      "example": "test diary"
                    },
                    "content": {
                      "type": "string",
                      "example": "this's content"
                    },
                    "images": {
                      "type": "array",
                      "items": {
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "5769097529f87b5c271728e5"
                          },
                          "image": {
                            "type": "string",
                            "example": ""
                          },
                          "description": {
                            "type": "string",
                            "example": "test description"
                          }
                        }
                      }
                    },
                    "detail_diary": {
                      "type": "array",
                      "items": {
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "5769097529f87b5c271728e5"
                          },
                          "date": {
                            "type": "date-time",
                            "example": "2016-10-05T04:27:58.144Z"
                          },
                          "title": {
                            "type": "string",
                            "example": "This's title"
                          },
                          "content": {
                            "type": "string",
                            "example": "This's content"
                          }
                        }
                      }
                    },
                    "permission": {
                      "type": "number",
                      "example": 1
                    },
                    "type": {
                      "type": "number",
                      "example": 1
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}